More improvments and fixes to signal handler

+++ src/include/debug.h
@@ -30,7 +30,7 @@ extern uintptr_t   trace_start, trace_end;
 extern box86context_t *my_context; // global context
 extern char* trace_func;
 extern uintptr_t fmod_smc_start, fmod_smc_end; // to handle libfmod (from Unreal) SMC (self modifying code)
-extern uint16_t default_fs;
+extern uint32_t default_fs;
 extern int jit_gdb; // launch gdb when a segfault is trapped
 #define LOG_NONE 0
 #define LOG_INFO 1
